refactor(cli): 统一stdio辅助函数导入路径风格
- 将cli.tsx和cli-args.ts中stdioHelpers的导入路径调整为统一的stdio-helpers格式
- 规范了模块文件命名，提高代码一致性
- 未改动核心功能逻辑，仅更改导入路径字符串
